move config helpers to separate file
